+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Added
 
 - Export a `getDefaultInitialViewState` function for getting a deafult inital view state given a loader and desired size of view.
+- Support interleaved RGB OME-TIFF files.
 - Added a README file for Avivator at `avivator/README.md`.
 
 ### Changed